cmd/cu: adds log command to show the log an env
